Handle messages that are binary equal except for the signature
There is a possibility that a message is created that contains exactly the same payload 
but has a different signature.                                                          
                                                                                        
This can occur when a message is created, forwarded, and for some reason the database is
reset.  The next time that the client starts the exact same message may be generated.   
However, because EC sigantures contain a random element the signature will be different.
                                                                                        
This results in continues transfers because the bloom filters identify the two messages 
as different while the community/member/global_time triplet is the same.                
                                                                                        
To solve this, we will silently replace one message with the other.  We choose to keep  
the message with the highest binary value while destroying the one with the lower binary
value.                                                                                  
                                                                                        
To further optimize, we will add both messages to our bloom filter whenever we detect   
this problem.  This will ensure that we do not needlessly receive the 'invalid' message 
until the bloom filter is synced with the database again.                               
                                                                                        
Returns False when the message is not a duplicate with anything in the database.        
Otherwise returns True when the message is a duplicate and must be dropped.             


git-svn-id: https://svn.tribler.org/abc/branches/release-5.3.x@21674 001aeff7-3401-0410-a489-f7902fc005dd