- AddressDestination is now CandidateDestination - The address property has been replaces with the candidate property (in Messages) - Basically all 'address' (host, port) tuples have been replaced with the Candidate object
The reason for this is that I need to be able to respond to incoming                      
messages, wherever they came from.  And because Dispersy is supposed                      
to interact with Swift we need to be able to respond whenever                             
something happend on the Swift network, but we can not guarantee that                     
we have (or are able to get) a Dispersy connection with the same node.                    
                                                                                          
Hence when we respond to something on the Swift network we need to be                     
able to respond by sending something to the Swift network, and not to                     
some address that may or may not have an available Dispersy listen                        
socket behind it.  Using Candidates allows this as we can make a                          
candidate that, instead of pointing to an ipv4 address, points to a                       
Swift socket.                                                                             
                                                                                          
Anyway, long story short: (host, port) is now Candidate.                                  


git-svn-id: https://svn.tribler.org/abc/branches/mainbranch@23493 001aeff7-3401-0410-a489-f7902fc005dd