Fixes magnetlink unittest.
The response protocol changed, instead of responding with data in a
'data' field inside the bencoded dict, the data is placed behind the
bencoded dict.

Added a sloppy_bdecode to bencode.py to allow us to get the offset
where the bencoded dict ends and the data begins.

Also, libtorrent does not reject duplicate requests.  Instead it gives
a normal response and outputs the following warning:

        LibtorrentDownloadImpl: alert performance_alert with message
        file.wmv: performance warning: max outstanding piece requests
        reached

Finally, variables named responce were renamed to response.