Fixed bugs when removing torrents from playlists
When removing torrents from a playlist, there were three errors. First,
the GUI was not immediately notified of the removal. Second,
there would be Dispersy warnings if we tried to remove the torrent from
a playlist a second time (while the first removal request was still
ongoing). Third, there was a error in the sql query that caused that the
right torrent to be removed was not always returned.