GigaChannel Community gossip caching
This makes GigaChannel Community cache the results of the SQL query
that selects the channel contents to gossip around. The resulting
blob is saved for reuse in the Community object. The blob will be sent
to 30 peers, and then the new query will be performed. This
significantly reduces the background DB activity.