Check timestamp of entry against the channel ver
Timestamps of incoming GigaChannel payloads are now checked
against the parent channel's local_version. This allows
us to stop getting old deleted results from previous channel
versions.
Also, fixed a severe bug in channel commit procedure. When committing
the personal channel, mdblob files should only contain entries with
timestamps <= to the mdblob filename, but the last mdblob must have
the same timestamp as the channel entry itself. Before, this was only
true if the channel entry was never edited (e.g. renamed) before commit.
Now we assign filenames based on the timestamp of the last entry in the
mdblob, except for the last file. For this, we create a new timestamp,
and then assign it to _both_ the filename and the updated channel entry.