Clean channel torrents with wrong names at startup
On startup, we now check if channel torrents dir names and torrent names in
.state files are compliant with our current naming scheme and remove the
rest. This is necessary to clean channel torrents from old betas
that still used pubkey-only dir names.

Also, to completely break with the earlier betas disk format, we
change the GigaChannel community id and database version.