Removed popular torrents explanation box
There were two complications when implementing this. First, I was unable to place the explanation box directly to the right of the text (as I did on the trust page). This is because the title widget is used for the channel navigation breadcrumb and modifying the width constraints breaks the breadcrumb computations. So I decided to put the widget to the right of the page.

Second, it seems that there is a bug that prevents customization of the QToolTip using CSS. The proper way to address this issue is by providing a custom mask but that's more complicated for now. The tooltip is not clipped correctly and there are some ugly borders visible. Fortunately, they blend with the background so they are hard to spot.