Modified shutdown procedure in GUI tests
In this PR, I slightly modify the shutdown procedure in the GUI tests. I dived a bit into the shutdown procedure when I triggered a segmentation fault in Qt when using dynamic row heights for tags (by overriding the sizeHint method). This segmentation fault would trigger in ~10% of the test runs, and by investigating the (fortunately, symbolicated!) stack trace, it seems to happen when Qt tries to do a final render operation when closing the Qt application. While a more effective solution is probably out of reach for now, it seems that we can workaround this by more gracefully shutting down the components in the GUI tests. I also believe this is a more reliable approach to unload the GUI tests.

Specifically, this PR ensures that we move to the 'shutting down' screen in the GUI when the GUI tests are done (and a screenshot is also made). Then, we shutdown the core. Only when the core process is finished, we close the Qt application.

Also, by making the `window` fixture dependent on the `tribler_api` fixture, I was able to remove the `tribler_api` fixture declaration from the test methods.