Added small delay after mouseclick in GUI test
It can take some time for the GUI to remove a tag after clicking the cross button, and this action is not instantly. Therefore, if the machine is busy, the tag could not have been removed correctly and the assert will fail. Adding a small timeout between the mouseclick and the assert should address this issue.

As much as I dislike hard-coded timeouts in tests, I believe this one is reasonable and necessary since Qt events are processed asynchronously.