Add test for `_filter_characters` and other minor fixes
Added a test for the new function `_filter_characters`.

Patched `test_get_name_as_unicode`. Before, it depended on manual maintenance of the right amount of `"?"` characters. Now, it is calculated programmatically. This ensures if the length changes in the future, the test doesn't break.

Minor coding style fixes.

Fixed pylint warning. This is needed so pylint doesn't complain for tests of the functionality of specific protected methods.