Refactor TorrentInfoEndpoint to fix JSON encoding issue
The code changes in this commit refactor the TorrentInfoEndpoint class in order to fix a JSON encoding issue. The previous implementation used json.dumps, which garbled binary data in the 'pieces' field. This commit removes that line of code and ensures that the GUI does not rely on this field.